% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{AlphamvecOpt}
\alias{AlphamvecOpt}
\title{AlphamvecOpt}
\usage{
AlphamvecOpt(K, z, alpha, mvec, niter)
}
\arguments{
\item{K}{total number of topics specified by the user}

\item{z}{current state of the assignment of topics}

\item{alpha}{Dirichlet concentration prior for topic distribution}

\item{mvec}{Dirichlet base prior for topic distribution}

\item{niter}{number of iterations to perfom}
}
\value{
The optimized value of the vector (= alpha*mvec)
}
\description{
Optimize the hyperparmeter vector (= alpha*mvec) given the current assignment of interaction patterns and topics
}
