% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GenerateDocs}
\alias{GenerateDocs}
\title{GenerateDocs}
\usage{
GenerateDocs(D, node, vocab, nIP, K, n.d, alpha, mvec, beta, b, eta, delta,
  sigma2_tau, l, support, netstat, timestat, base.edge = NULL,
  base.text = NULL, topic_token_assignments = NULL, backward = FALSE,
  base = FALSE)
}
\arguments{
\item{D}{number of documents to be generated}

\item{node}{vector of node id's (ID starting from 1)}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{n.d}{number of words in a document (fixed constant for now)}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{b}{coefficients for recipients}

\item{eta}{coefficients for timestamps}

\item{delta}{tuning parameter for the number of recipients}

\item{sigma2_tau}{variance parameter for the timestamps}

\item{l}{topic-interaction pattern assignment}

\item{support}{support of latent recipients}

\item{netstat}{which type of network statistics to use ("intercept", "dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("timeofday", "dayofweek")}

\item{base.edge}{edges before 384 hours that is used to calculate initial history of interactions}

\item{base.text}{texts corresponding to base.edge}

\item{topic_token_assignments}{matrix of topic-token assignments}

\item{backward}{Logigal indicating whether we are generating backward samples (if FALSE -> forward)}

\item{base}{Logical indicating whether or not we are generating base edges (< 384)}
}
\value{
generated data including (author, recipients, timestamp, words)
}
\description{
Generate a collection of documents according to the generative process of IPTM
}
