% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{GiR_stats}
\alias{GiR_stats}
\title{GiR_stats}
\usage{
GiR_stats(GiR_sample)
}
\arguments{
\item{GiR_sample}{one sample from generative process}
}
\value{
A vector of statistics calculated from one GiR sample
}
\description{
Calculate several statistics from samples generated from forward or backward sampling
}
