% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Schein}
\alias{Schein}
\title{Schein}
\usage{
Schein(Nsamp, D, node, vocab, nIP, K, n.d, alpha, mvec, beta, prior.b,
  prior.delta, prior.eta, prior.tau, sigma.Q, Outer, Inner, burn, thin,
  netstat = c("dyadic"), timestat = c("timeofday", "dayofweek"), base.edge,
  base.text, generate_PP_plots = TRUE)
}
\arguments{
\item{Nsamp}{number of GiR samples to be generated}

\item{D}{number of documents to be generated per each sample}

\item{node}{vector of node id's (ID starting from 1)}

\item{vocab}{all vocabularies used over the corpus}

\item{nIP}{total number of interaction patterns}

\item{K}{total number of topics}

\item{n.d}{number of words in a document (fixed constant for now)}

\item{alpha}{Dirichlet concentration prior for document-topic distribution}

\item{mvec}{Dirichlet base prior for document-topic distribution}

\item{beta}{Dirichlet concentration prior for topic-word distribution}

\item{prior.b}{prior mean and covariance of b in multivariate Normal distribution}

\item{prior.delta}{prior mean and variance of delta in Normal distribution}

\item{prior.eta}{prior mean and covariance of eta in multivariate Normal distribution}

\item{prior.tau}{prior shape and scale parameter of sigma2_tau in inverse-Gamma distribution}

\item{sigma.Q}{proposal distribution variance parameter}

\item{Outer}{size of outer iterations}

\item{Inner}{size of inner iteration for Metropolis-Hastings updates}

\item{burn}{iterations to be discarded at the beginning of Metropolis-Hastings chains}

\item{thin}{the thinning interval of Metropolis-Hastings chains}

\item{netstat}{which type of network statistics to use ("dyadic", "triadic", "degree")}

\item{timestat}{additional statistics to be used for timestamps other than netstat ("timeofday", "dayofweek")}

\item{base.edge}{artificial collection of documents to be used as initial state of history}

\item{base.text}{artificial collection of documents to be used as initial state of history}

\item{generate_PP_plots}{Logical indicating whether to draw PP plots}
}
\value{
Forward and Backward samples and corresponding test results
}
\description{
Schein (preliminary GiR) test for the IPTM
}
